requirements.txt
fastapi==0.95.1
requests==2.31.0
uvicorn==0.22.0
motor==3.1.2
python-multipart==0.0.6
python-dateutil==2.8.2
python-slugify==8.0.1
aiofiles==23.1.0
email-validator==2.0.0.post2
aiohttp==3.8.4
nest-asyncio==1.5.6
httpx==0.24.1
python-jose==3.3.0
cryptography==40.0.2
passlib==1.7.4
bcrypt==4.0.1
redis==4.5.5
elasticsearch==8.8.2